+++ apps/passport-client/pages/index.tsx
@@ -31,19 +31,17 @@ import {
 import { addDefaultSubscriptions } from "../src/defaultSubscriptions";
 import {
   Action,
+  dispatch,
   StateContext,
-  StateContextState,
-  dispatch
+  StateContextState
 } from "../src/dispatch";
 import { Emitter } from "../src/emitter";
 import {
-  loadAnotherDeviceChangedPassword,
   loadEncryptionKey,
   loadIdentity,
   loadPCDs,
   loadSelf,
   loadSubscriptions,
-  loadUserInvalid,
   saveIdentity,
   saveSubscriptions
 } from "../src/localstorage";
@@ -187,9 +185,6 @@ async function loadInitialState(): Promise<AppState> {
   const self = loadSelf();
   const pcds = await loadPCDs();
   const encryptionKey = loadEncryptionKey();
-
-  const userInvalid = loadUserInvalid();
-  const anotherDeviceChangedPassword = loadAnotherDeviceChangedPassword();
   const subscriptions = await loadSubscriptions();
 
   subscriptions.updatedEmitter.listen(() => saveSubscriptions(subscriptions));
@@ -200,9 +195,7 @@ async function loadInitialState(): Promise<AppState> {
 
   let modal = "" as AppState["modal"];
 
-  if (userInvalid) {
-    modal = "invalid-participant";
-  } else if (
+  if (
     // If on Zupass legacy login, ask user to set passwrod
     self != null &&
     self.salt == null
@@ -217,8 +210,6 @@ async function loadInitialState(): Promise<AppState> {
     pcds,
     identity,
     modal,
-    userInvalid: userInvalid,
-    anotherDeviceChangedPassword,
     subscriptions,
     resolvingSubscriptionId: undefined
   };