+++ packages/rsa-pcd/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pcd/rsa-pcd
 
+## 0.2.1
+
+### Patch Changes
+
+- new package versions.
+- Updated dependencies
+  - @pcd/passport-ui@0.7.1
+  - @pcd/pcd-types@0.7.1
+  - @pcd/util@0.1.1
+
 ## 0.2.0
 
 ### Minor Changes